Rename polymer/behavior-descriptor.ts to behavior.ts
