+++ src/polymer/behavior-scanner.ts
@@ -21,7 +21,7 @@ import {JavaScriptDocument} from '../javascript/javascript-document';
 import {JavaScriptScanner} from '../javascript/javascript-scanner';
 import * as jsdoc from '../javascript/jsdoc';
 
-import {ScannedBehavior} from './behavior-descriptor';
+import {ScannedBehavior} from './behavior';
 import {declarationPropertyHandlers, PropertyHandlers} from './declaration-property-handlers';
 import * as docs from './docs';
 import {toScannedPolymerProperty} from './js-utils';
