+++ src/polymer/docs.ts
@@ -15,7 +15,7 @@
 import * as jsdoc from '../javascript/jsdoc';
 import {ScannedEvent, ScannedFeature, ScannedProperty} from '../model/model';
 
-import {ScannedBehavior} from './behavior-descriptor';
+import {ScannedBehavior} from './behavior';
 import {ScannedPolymerCoreFeature} from './feature-descriptor';
 import {ScannedFunction, ScannedPolymerElement, ScannedPolymerProperty} from './polymer-element';
